Create Python Program for focal length of a spherical mirror
Focal length is the distance between the center of the mirror to the principal foci. In order to determine the focal length of a spherical mirror we should know the radius of curvature of that mirror. The distance from the vertex to the center of curvature is called radius of curvature.

The focal length is half the radius of curvature.