Create Third Umpire Decision Review System (DRS Gully Cricket) | Python
In today’s tutorial, we are going to create a “Third Umpire Decision Review System”

One of the most treasured memories of growing up is playing cricket with friends. Just like me, there are countless other people whose favourite part of the day is to grab a pair of bat and ball and just play. Gully cricket is the most popular form of sport, is played in everywhere. So, today’s project is going to be very helpful for those cricket fans who want to create their own decision review system. 

Prerequisite: The prerequisite of this project is the basic knowledge of python.

The third empire takes the decision about no-ball, runout, or catch out. Here we are going to make the decision review system using python which will take an accurate decision about whether the batsman is out or not out.